requirements.txt
asgiref==3.8.1
Django==5.1.2
django-grappelli==4.0.1
django-taggit==6.1.0
djangorestframework==3.12.4
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2